+++ packages/util-actor/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 - YYYY-MM-DD **BREAKING?** -- description
 
+### 0.0.2 -- 2020-08-13
+
+- 2020-08-13 -- Added exported fn `subscriber(address, spawnFn, callback)`
+- 2020-08-13 -- Added exported fn `snapshoter(address, spawnFn)`
+- 2020-08-13 -- Added `ctx.sendSelf()`
+- 2020-08-13 -- Added `ctx.subscriberCount()`
+- 2020-08-13 -- Added `ctx.hasSubs()`
+
 ### 0.0.1 -- 2020-07-17
 
 - 2020-07-17 -- Initial Implementation